Add userCurrentUser hook
